requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
Appium-Python-Client==2.7.1
async-generator==1.10
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
coloredlogs==15.0.1
colorlog==6.7.0
configparser==5.3.0
coverage==7.0.5
cryptography==38.0.4
Cython==0.29.32
dl==0.1.0
docutils==0.19
et-xmlfile==1.1.0
execnet==1.9.0
h11==0.14.0
HTMLParser==0.0.2
humanfriendly==10.0
idna==3.4
importlib-metadata==5.1.0
iniconfig==2.0.0
ipaddr==2.2.0
jaraco.classes==3.2.3
keyring==23.11.0
lockfile==0.12.2
lxml==4.9.1
more-itertools==9.0.0
mypy-extensions==0.4.3
numpy==1.23.5
openpyxl==3.0.10
ordereddict==1.1
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
protobuf==4.21.10
py==1.11.0
pycparser==2.21
pyOpenSSL==22.1.0
PySocks==1.7.1
pytest==7.2.0
pytest-cov==4.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-xdist==3.1.0
selenium==4.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
zipp==3.11.0

pip~=22.3.1
argparse~=1.4.0
Pillow~=10.0.0
setuptools~=65.5.0
platformdirs~=3.0.0
requests~=2.31.0